Restructure project: extract models, separate bot/web, fix bugs
- Extract SlackMessage, User, Job into slackbot/models/ package
- Move SlackBot class from __init__.py to bot.py, re-export from __init__
- Create web.py with Flask app factory (create_app)
- Fix handle_command not breaking after first plugin match
- Fix load_plugin/unload_plugin returning None instead of success/failure
- Fix request.data mutation in event handler, use get_json() instead
- Fix UserManager.save() to sync data reference before writing
- Make DataManager writes atomic via tempfile + os.replace
- Add dirty tracking to DataManager to skip redundant writes

+++ slackbot/managers/data.py
@@ -1,25 +1,41 @@
 import os
 import json
+import tempfile
+
 
 class DataManager:
     def __init__(self, bot, file, default=None):
         self.bot = bot
         self.file_name = file
-        self.bot.log.info("setting up datamanager for '%s'" % self.file_name)
-        self.file = os.path.join(self.bot.state_file_base, '%s.json' % self.file_name)
+        self.bot.log.info(f"setting up datamanager for '{self.file_name}'")
+        self.file = os.path.join(self.bot.state_file_base, f'{self.file_name}.json')
         self.data = default or {}
+        self._dirty = False
 
-        self.bot.log.info("loading data file '%s'" % self.file_name)
         if os.path.isfile(self.file):
+            self.bot.log.info(f"loading data for '{self.file_name}'")
             with open(self.file) as f:
-                self.bot.log.info("loading data for '%s'" % self.file_name)
                 self.data = {**self.data, **json.load(f)}
+
+    def mark_dirty(self):
+        self._dirty = True
+
+    def save(self, force=False):
+        if not force and not self._dirty:
+            return
+        self.bot.log.info(f"saving data file '{self.file_name}'")
+        dir_name = os.path.dirname(self.file)
+        fd, tmp_path = tempfile.mkstemp(dir=dir_name, suffix='.tmp')
+        try:
+            with os.fdopen(fd, 'w') as f:
+                json.dump(self.data, f)
+            os.replace(tmp_path, self.file)
+        except BaseException:
+            os.unlink(tmp_path)
+            raise
+        self._dirty = False
 
     def set(self, data):
         self.data = data
-        self.save()
-
-    def save(self):
-        self.bot.log.info("saving data file '%s'" % self.file_name)
-        with open(self.file, 'w') as f:
-            json.dump(self.data, f)
+        self._dirty = True
+        self.save(force=True)
